Rename PlayGame and Game
